Add details to simplified producer
