+++ pubsub-mapped-api/pom.xml
@@ -14,6 +14,11 @@
       <version>0.2-EXPERIMENTAL</version>
     </dependency>
     <dependency>
+      <groupId>com.google.cloud</groupId>
+      <artifactId>google-cloud-pubsub</artifactId>
+      <version>0.9.2-alpha</version>
+    </dependency>
+    <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>4.12</version>
