+++ pubsub-mapped-api/src/main/java/com/google/pubsub/clients/producer/PubsubProducerConfig.java
@@ -43,6 +43,9 @@ public class PubsubProducerConfig extends AbstractConfig {
   public static final String PROJECT_CONFIG = "project";
   private static final String PROJECT_DOC = "GCP project to use with the publisher.";
 
+  public static final String MAX_REQUEST_SIZE_CONFIG = "max.request.size";
+  private static final String MAX_REQUEST_SIZE_DOC = "The maximum size of a request in bytes.";
+
   static {
     CONFIG =
         new ConfigDef()
@@ -52,7 +55,8 @@ public class PubsubProducerConfig extends AbstractConfig {
                 VALUE_SERIALIZER_CLASS_CONFIG, Type.CLASS, Importance.HIGH, VALUE_SERIALIZER_CLASS_DOC)
             .define(BATCH_SIZE_CONFIG, Type.INT, 1, Importance.MEDIUM, BATCH_SIZE_DOC)
             .define(ACKS_CONFIG, Type.STRING, "1", Importance.MEDIUM, ACKS_DOC)
-            .define(PROJECT_CONFIG, Type.STRING, Importance.HIGH, PROJECT_DOC);
+            .define(PROJECT_CONFIG, Type.STRING, Importance.HIGH, PROJECT_DOC)
+            .define(MAX_REQUEST_SIZE_CONFIG, Type.INT, 1*1024*1024, atLeast(0), Importance.MEDIUM, MAX_REQUEST_SIZE_DOC);
   }
 
   PubsubProducerConfig(Map<?, ?> properties) {
