+++ pubsub-mapped-api/src/main/java/com/google/pubsub/clients/producer/internals/PubsubFutureRecordMetadata.java
@@ -1,0 +1,38 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+/*package com.google.pubsub.clients.producer.internals;
+
+private static class PubsubFutureRecordMetadata implements Future<RecordMetadata> {
+  public boolean cancel(boolean b) {
+    return false;
+  }
+
+  public boolean isCancelled() {
+    return false;
+  }
+
+  public boolean isDone() {
+    return false;
+  }
+
+  public RecordMetadata get() throws InterruptedException, ExecutionException {
+    return null;
+  }
+
+  public RecordMetadata get(long l, TimeUnit timeUnit) throws InterruptedException, ExecutionException, TimeoutException {
+    return null;
+  }
+}*/
